Reorganized codebase, added privacy banner and user session tracking
In this update, we restructured our CSS and JS files for better organization and maintainability hence, the renaming and moving of certain files and resources.

Additionally, we added a new privacy banner in the main index.html file to inform users about our data usage and privacy policies. Specific CSS styling is also added for the new privacy banner.

In the Python back-end, we modified the Flask application to keep track of whether a user has submitted data or not, using the Flask session interface for this purpose. This change was necessary to appropriately redirect the users based on their interaction with the site.

To support these changes, new HTML templates for a privacy page and a 404 error page were created, extending the interface and improving the user experience.

Overall, this commit improves the site structure, user experience, and data handling.