Replace hardcoded secret key with environment variable reference
This commit removes a hardcoded secret key in the Flask app setup and instead replaces it with a reference to pull the value from environment variables. This change is part of an effort to follow best security practices, where sensitive data must not be hardcoded into the script and should instead be securely held as environment variables.