Update Privacy Policy text and add .gitignore
This commit updates the Privacy Policy text on 'templates/privacy.html'
and adds a new '.gitignore' file.

The privacy section has been revised for clarity and specificity. We've
changed the last update date to 14th September 2023 and unambiguous
text on the use of cookies has been added to ensure users that the website
only tracks their provided information and not their personal data.

A new '.gitignore' file has been added to the repository. It currently
ignores .env files, which is crucial for preventing sensitive environment
variable information from being accidentally committed to the repository.