Replace unnecessary logs with ajax calls; Refactor data parsing; Add rightContainer.html
Implemented ajax calls to replace console logs in `button-listener.js` and enhanced the performance of the app by refactoring the data parsing function in `get_json.py`. This also involved adding a new `data-info` attribute to the buttons for better manipulation of item data and parsing this data into a server-side route. A new static file, `rightContainer.html`, was added to dynamically replace data in the right container. The `success.css` file was also adjusted to incorporate these changes.