Add "Search" button functionality and styles
In this commit, a new "Search" button has been introduced to extend the user interactivity. The button, named as "button3", has been added to the HTML with corresponding CSS styles for consistent appearance. The JavaScript logic has also been updated to manage the visibility and behavior of the new button. When "button3" is pressed, the other buttons are hidden which imitates a switch-like functionality. Some modifications have been made in app.py and rightContainer.html files to accommodate these changes while also improving code readability.